app/requirements.txt
fastapi
pydantic>=2.0.0,<3.0.0
uvicorn[standard]
asyncpg
psycopg2-binary
python-dotenv
alembic
pydantic-settings
azure-storage-blob
aiohttp
aiofiles
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytz
python-dateutil
azure-cognitiveservices-speech
pygame
openai
tiktoken
numpy
soundfile
